Fix WIFI_SERVICE-related memory leak
From lintVital: the WIFI_SERVICE must be looked up on the Application
context or memory will leak on devices < Android N.

On versions prior to Android N (24), initializing the WifiManager via
Context#getSystemService can cause a memory leak if the context is not
the application context. Change context.getSystemService(...) to
context.getApplicationContext().getSystemService(...).